+++ cider/exec/nextgen/CMakeLists.txt
@@ -26,4 +26,8 @@ if(ENABLE_ASN)
 endif()
 
 add_subdirectory(jitlib)
+<<<<<<< HEAD
 add_subdirectory(translator)
+=======
+# add_subdirectory(operators)
+>>>>>>> add9bf2 (Staged)
